Move test-utils to tests/utils because it is  only used for internal testing
