+++ tests/__tests__/import.spec.ts
@@ -1,5 +1,5 @@
 import runJest from '../__helpers__/runJest';
-import { printStdStreams } from '../../src/test-utils';
+import { printStdStreams } from '../utils/test-utils';
 
 describe('import with relative and absolute paths', () => {
   it('should run successfully', () => {
